fixed jfr/event/sampling/TestNative test
This test was brouht in by 8223147: JFR Backport and never passed.
Despite not being wrong nor having jfr broken in jdk8.
jdk8 testsuite simply lacks native support casses.

This is adding the shell wrapper which is compiling and properly setting up
the native longSleep  in native shared object.

This comit was tested only on linux.
After this commit whole jdk_jfr group passes
Would be nice if original authors egahlin and mgronlun doublechekced if
the test is still testing  the same