Backport 791b427f4410057cdcdf8fd8ea0dcce71f7dc513
When the docker serivice work normally on the test machine, this test
will always fail. This test want to verify the API
jdk.internal.platform.Metrics.systemMetrics().getMemoryFailCount() work
normally or not. The API return memory allocate fail times in jvm. But,
before this PR, everytime it allocate 8M memory, the API has no chance
the catch the memory allocate fail, the jvm was killed by OOM. Change 8M
increments to 1K mean to avoid OOM killed for the jvm in docker
container.

Signed-off-by: sendaoYan <yansendao.ysd@alibaba-inc.com>